Organized Fastify a bit
(Now that I know a bit more, this is pretty messy, but I don't want to break anything by changing the actual endpoints)

Also changed version to v0.12.3.
