+++ backend/src/routes/api.ts
@@ -1,0 +1,20 @@
+import { FastifyPluginAsync } from 'fastify';
+import upload from './upload';
+import avatarUploadRoute from './avatars';
+import statsRoute from './stats';
+import postDeleteRoute from './delete/posts';
+import avatarDeleteRoute from './delete/avatars';
+import postReplaceRoute from './replace';
+import integrityCheck from './checks';
+
+const apiRoutes: FastifyPluginAsync = async (fastify) => {
+  fastify.register(upload, { prefix: '/api' });
+  fastify.register(avatarUploadRoute, { prefix: '/api' });
+  fastify.register(statsRoute, { prefix: '/api' });
+  fastify.register(postDeleteRoute, { prefix: '/api' });
+  fastify.register(avatarDeleteRoute, { prefix: '/api' });
+  fastify.register(postReplaceRoute, { prefix: '/api' });
+  fastify.register(integrityCheck, { prefix: '/api' });
+};
+
+export default apiRoutes;
