+++ frontend/src/core/constants/version.ts
@@ -2,5 +2,5 @@
 // Manually changing this yourself could lead to accidental file deletion.
 
 // Only developers should update this when pushing a new version.
-export const CURRENT_VERSION = '0.12.2';
-export const LEGIBLE_VERSION = 'pre-v0.12.2';
+export const CURRENT_VERSION = '0.12.3';
+export const LEGIBLE_VERSION = 'pre-v0.12.3';
